requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
PyPDF2==3.0.1
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
